requirements.txt
requests
re
sys
signal
pwntools
